Add new functionality and some kind of backwards compatibility
Add compatibility with the previous Beta version of the api (use auth_token & api_key)
Fix update method for file.
Add logic to make the name unique in case of a name conflict.
Add method to get the collaborations for a folder.
Get all items when info is requested about a folder.